Space background sizes was giving me issues on display so I re-cropped the set from the original.
I added scaling for oversize images.  Any set of three images can be used for the background as long as the images in the set have the same width and length of each other.

Player initializes to middle of screen bottom. Enemy initializes to center of display and performs box patrol.

-R